Crate cjdns_keys

CJDNS keys types.

Structs

CJDNSKeys
Convenience type for managing all CJDNS key types in one variable.

CJDNSKeysApi
Type that encapsulates some crate functions making it safer for its users to work with randomly created keys.

CJDNSPrivateKey
CJDNS private key type

CJDNSPublicKey
CJDNS public key type

CJDNS_IP6
CJDNS IP6 type

Type Definitions

IpV6
CJDNS IPv6 type

PrivateKey
CJDNS private key type

PublicKey
CJDNS public key type